Add files via upload
test.py was an initial test of the bot, the current build is far different; keep in mind that this bot had 3 different purposes throughout its time in existence. Each idea also had multiple builds. The final build is in place because it simply works best with the twitter API, not because it's my favorite.